small details
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@304 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CWorm_AI.cpp
@@ -2230,7 +2230,7 @@ bool CWorm::weaponCanHit(int gravity, float speed, CVec cTrgPos, CMap *pcMap)
 	float cos_alpha = cos(alpha);
 	float tan_alpha = tan(alpha);
 	y = 0;
-	float dy;
+	int dy;
 	
 	if (max_x == 0) {
 		if(speed && fabs(cos_alpha)>0.1f) return false;
